Make Chakra UI (and its peer deps) peer dependencies
This commit makes `@chakra-ui/react`, `@emotion/styled`, `@emotion/react` and `framer-motion` peer dependencies of the `@vygruppen/spor-react` package.

This means you'll have to install these dependencies yourself when you're adding the `@vygruppen/spor-react` library as a dependency. It's a bit more to copy paste in, but also a lot more flexible than the alternative.