Change the way we specify our theme
Instead of blindly adding to the default theme, enabling tons of colors and sizes that aren't meant for use in our component library, we create our own. If we haven't specified a component yet, we fall back to the default.

This might break a few default components (missing colors etc), but that will be fixed as we start implementing those components