Add line tag components for Elm (#429)
Background

We want to implement the new design for line tags in Elm. Figma: https://www.figma.com/file/VTQxSbM0vwGfS5Zz0vLJG9/Linjetags

The two main types are the InfoTag and the TravelTag. The InfoTag is a simpler variant than the TravelTag, and does not have any background color. The implementation is based on the already implemented components in React Native.

Important note:

Our API model in Itinerary exposes a Line object with colors for lines. Today, these are used in both our existing line tags, for lines in the map and for borders inside the travel details.

There is no good way in the data from Entur to separate between train types such as regional trains and regional express trains, etc.. their submode is not that granular. The only thing that really separates them from each other, is their line ID. Since we already use colors from the Line object directly (and have a mapping for colors in backend), it feels a bit overkill to add yet another field to separate train types into further submodes, and then map them to the correct line tag in our front end. We also already have other fields that can help us in the mapping of which variant to use (mode, operator, submode). The train types also use the same icon, so the only thing that differs between them, is the color.

If this "train type" field is something we need later, we can look into adding it then.

My solution is therefore to have the default components use the colors defined by the variant, but also being able to send in color and bakcgroundColor as optional properties, that override the default. The backend mapping will ensure that these colors are correct. This way, it will be possible to use the component both in a context with such a Line object and in other where we might not have that.

Solution

Implement the following components (using the builder pattern):

TravelTag
InfoTag
LineTagText
LineTagIcon
By implementing each of these as their own components,, it will be easier for consumers to use each of these more flexible.