Add workaround for click throughs on mobile devices in CalendarCell and ListBox (#799)
## Background

We experience issues with click throughs in the DatePicker's calendar and in the ComboBox's list box on mobile devices. This was also noticed in React Aria's own examples.

## Solution

The issue was reported to react-spectrum, and they already linked to other related issues that had been reported before. In one of these, a workaround was suggested. 

The workaround is now tested locally in Spor, and seems to work great (and one can still scroll inside the list box). I believe we can release a new version and test this out in our app (otherwise I would have to copy the components to the app and test them there, which is a bit of a hassle, but not impossible, of course).