Change type of calendar trigger button to "button" (#807)
## Background

When not using type "button" on a button, the default is "submit". Somehow, the button does not like that we use the `onPress` from the `buttonProps`, and Enter or Space buttons do not work if the type is set to "button".

## Solution

Change the type of the button to "button" to make sure it does not submit anything. Also, make sure that the button still can be opened using "Enter" and "Space".