Override minimum left padding in ComboBox and InfoSelect popovers (#811)
## Background

A negative margin was added to the popover due to that the ComboBox popover was displaced some pixels to the right in the app it was used after removing the overlay. The popover will in fact _always_ be displaced to the right relative to the input field when there is less than 12 px padding to the left in the container. This is because the [default minimum container padding is 12 px](https://github.com/adobe/react-spectrum/discussions/5009).

By adding a negative margin, it worked against its purpose for InfoSelect and ComboBox when there is enough left padding, causing the popover to be displaced the opposite way.

## Solution

We can override the minimum padding by passing `containerPadding: 0` to `usePopover`.