Added type declaration for chakra
There is a weird bug when you have moduleResolution set to bundler, causing modules that use a more direct import style to become unresolvable. The solution for this wasn’t obvious.

The reason turned out to be related to our spor-react library, where we use Chakra’s forwardRef. The definition is specified to be within some dist folder in Chakra.

I suppose there might be some tricks to work around this by importing it differently, but it has become too time-consuming to chase this any further, especially since we might switch to another component library later on.

The problem doesn’t occur if you set moduleResolution to node.