Update PhoneNumberInput name prop (#1576)
## Background

Not being able to set specific `name`-attributes for the country-code
and national-number inputs is a bit problematic when using
[Conform](https://conform.guide/).

Conform expects a certain naming convention (e.g.
`phoneNumber.countryCode`) which is currently not possible to achieve as
the PhoneNumberInput adds a suffix to the name, i.e.
`${name}-country-code` and `${name}-phone-number`.
This results in relying on hidden inputs and other hacky solutions when
using PhoneNumberInput with Conform.

## Solution

Updated the name prop to

```
name?:
  | string
  | { countryCode: string; nationalNumber: string; };
```

If given a string it works the same way as previously, if given an
object no `-country-code`/`-phone-number` suffix is added.

## UU checks

- [ ] Sanity documentation has been / will be updated (if neccessary)

## How to test

Check that the PhoneNumberInput works as expected when given a string or
an object for the name.