Add prop for selecting which country codes you want (#1670)
## Background

For PhoneNumberInput it is unnecessary and confusing to show all
countryCodes. For my task I only need norwegian and swedish phone
numbers. Therefore I need a way to only show the relevant countryCodes.

## Solution

Added a prop called "allowedCountryCodes" used for filtering the country
code items.

## Chakra update checklist

( No checklists as this is only a single prop)

If no packages, only docs has been changed:

- [ ] Documentation version has been bumped (package.json in docs)

Everything about making a React component:
https://spor.vy.no/guides/how-to-make-new-react-components

HOW TO MAKE A CHANGESET:
Go here:
https://spor.vy.no/guides/how-to-make-new-react-components#creating-a-pr-and-publish-package

## How to test

Try to add ["+47", "+46"] to allowedCountryCode and check if it works.

## Screenshots

( See "How to test" and check visually)

---------

Co-authored-by: Eivind Yu Nilsen <eivindyunilsen@hotmail.com>