Patch lottie-web to work with newer versions of node (#1694)
## Background
The package `lottie-web` does not work properly with SSR on newer
versions of node (node > 20). This is because node has made changes in
their navigator api so it is now available on the server as well. This
causes a check `lottie-web` has to determine what context it is in to
not work as intended, leading to a `document is undefined` error.

Link to relevant github issue:
https://github.com/airbnb/lottie-web/issues/2739

## Solution

Create a patch version of the `lottie-web` package with `pnpm patch`. In
the patched version we update the check to see if the document is
present instead of the navigator.