New studio designmanual (#1752)
## Background

In order to join designmanual and spor togheter, we need a fresh new
sanity studio and new datasets. These are the changes allows to connect
to the new sanity project

## Solution

Create a new project into the organization and deploy a new empty
project locahost. Copy all the schemas from the current project to the
new one

## How to test

Move to apps/design-studio folder and run pnpm dev it opens a fresh
empty studio with same schemas as the current studio. You need sanity
permission to get into studio ask for it to assignees