fix(input): resolve missing label error when specifying id on inputs (#1846)
## Background
The id for a field is typically auto-generated. However, when an id is
explicitly specified (e.g., in an Input component), it is not correctly
passed to the Field component, causing the label to be improperly
linked.

## Solution
Ensure the id is properly passed to the Field component to establish the
correct connection with the label.