Portable text components (#1873)
<!---
Thanks for creating a Pull Request! 🎉

✅ Keep your PR as small as possible.
📘 React component guide:
https://spor.vy.no/guides/how-to-make-new-react-components
📦 How to make a changeset:
https://spor.vy.no/guides/how-to-make-new-react-components#creating-a-pr-and-publish-package
💡 Preferably use Conventional Commits:
https://www.conventionalcommits.org/en/v1.0.0/

This template serves as a guideline; feel free to remove sections that
do not apply to your pull request.
-->

## Background

The new design manual need to be updated by sanity for the documentation

<!-- Why this task is needed, context, and problem it solves -->

---

## Solution

Create a new Sanity project and reuse the components from vy.no project
to make the CMS able to create pages for the identity pages and make the
frontend documentation able to translate the portableText components
with the correct design

<!-- What has been done and why this approach was chosen -->

---

## General Checklist

- [ ] I have updated documentation if necessary
- [ ] I have verified the design aligns with the latest Figma sketches.
- [ ] I have created a changeset if publishing is required

## Accessibility checklist

For changes impacting the user interface or functionality, ensure the
following:

- [x] It is possible to use the keyboard to reach your changes
- [x] It is possible to enlarge the text 400% without losing
functionality
- [x] It works on both mobile and desktop
- [x] It works in both Chrome, Safari and Firefox
- [x] It works with VoiceOver
- [ ] There are no errors in aXe / SiteImprove-plugins / Wave

---

## Screenshots

| Before | After |
| ------ | ----- |
|        |       |

---------

Co-authored-by: Asgeir Hunshamar <asgeirhuns@gmail.com>