Add env variables on terraforn and load them on frontend (#1985)
<!---
Thanks for creating a Pull Request! 🎉

✅ Keep your PR as small as possible.
📘 React component guide:
https://spor.vy.no/guides/how-to-make-new-react-components
📦 How to make a changeset:
https://spor.vy.no/guides/how-to-make-new-react-components#creating-a-pr-and-publish-package
💡 Preferably use Conventional Commits:
https://www.conventionalcommits.org/en/v1.0.0/

🚀 Deploy to environments:
- Comment `.preview` to deploy to preview environment
- Comment `.deploy test` to deploy to test environment

This template serves as a guideline; feel free to remove sections that
do not apply to your pull request.
-->

## Background

Closes # <!-- Github issue number -->

<!-- Why this task is needed, context, and problem it solves -->

In order to use the presentation tool from sanity some environment
variables needed to be loaded from server to work on stage and prod

---

## Solution

<!-- What has been done and why this approach was chosen -->

Add variables on terraform and load on frontend

---------

Co-authored-by: Asgeir Hunshamar <asgeir.hunshamar@vy.no>