Fix select ssr portalled (#2026)
<!---
Thanks for creating a Pull Request! 🎉

✅ Keep your PR as small as possible.
📘 React component guide:
https://spor.vy.no/guides/how-to-make-new-react-components
📦 How to make a changeset:
https://spor.vy.no/guides/how-to-make-new-react-components#creating-a-pr-and-publish-package
💡 Preferably use Conventional Commits:
https://www.conventionalcommits.org/en/v1.0.0/

🚀 Deploy to environments:
- Comment `.preview` to deploy to preview environment
- Comment `.deploy test` to deploy to test environment

This template serves as a guideline; feel free to remove sections that
do not apply to your pull request.
-->

## Background

The Select component won't render properly when used in a WebView in the
app. The reason might be hydrating differences when rendering on server
side before presenting in the app.


<!-- Why this task is needed, context, and problem it solves -->

---

## Solution

This disables the Portal during SSR (when window is undefined) and
enables it on the client side, preventing the useLayoutEffect warnings
while maintaining the desired behavior in the browser.
<!-- What has been done and why this approach was chosen -->
Have tested locally that it works the same. Need to publish the change
to test if the problem is actually fixed in test/stage.

---

## General Checklist

- [x] I have created a changeset if publishing is required